fix(ci): upgrade pipewire crates for Ubuntu 22.04 compat
- Upgrade pipewire/libspa/libspa-sys from 0.7 to 0.8 to fix build
  failure caused by missing spa_type_video_flags symbol
- Add pipewire-upstream PPA to CI for newer PipeWire headers (0.3.65+)
- Add libspa-0.2-dev to CI dependencies
- Fix clippy warnings: redundant closures, unnecessary casts, manual
  strip_prefix, derivable Default impl, collapsible if statements
- Add #[allow(dead_code)] for unused but API-planned audio methods

+++ src-tauri/src/capture/linux/highlight.rs
@@ -74,8 +74,8 @@ fn get_hyprland_monitors() -> Vec<HyprMonitorInfo> {
         Ok(monitors) => {
             monitors.iter().map(|m| HyprMonitorInfo {
                 name: m.name.clone(),
-                x: m.x as i32,
-                y: m.y as i32,
+                x: m.x,
+                y: m.y,
                 width: m.width as i32,
                 height: m.height as i32,
                 scale: m.scale as f64,
